backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
google-auth-oauthlib==1.1.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
Werkzeug==2.3.0
